refactor: データソース層の重複解消とインターフェース統一
- 抽象インターフェースとDrift実装の冗長性を解消
- 3つのエンティティ(Store, Photo, VisitRecord)で同様の統合を実施
- 単一ファイルに抽象インターフェースと実装を統合
- クラス名をStoreLocalDatasourceImpl等に統一
- DIコンテナ設定を新しいクラス名に更新
- 全テストファイルを新しい構造に対応

影響範囲:
- lib/data/datasources/ の3つのローカルデータソース
- lib/core/di/ のDIコンテナ設定
- test/unit/data/datasources/ の対応テストファイル

期待効果:
- コードベースの簡素化
- 保守コストの削減
- 新しいデータソース追加時の効率化
